cheque_validation_project/requirements.txt
streamlit
pymongo
pytesseract
pdf2image
google-generativeai
python-dotenv
plotly
opencv-python
pillow
numpy